requirements.txt
discord
python-dotenv
psutil
Flask
Flask-Login
Flask-SQLAlchemy
requests